requirements.txt
monai==1.1.0
monailabel-weekly==0.5.dev2307
scikit-learn==1.2.1
pandas==1.5.3
matplotlib==3.7.0
ray==2.2.0
timm==0.6.12
torch-summary
ipykernel
ipywidgets